betterCorq-main/requirements.txt
Flask
requests
python-dotenv
pytz